削除機能の実行可能者の制限

+++ app/controllers/items_controller.rb
@@ -39,8 +39,10 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    item = Item.find(params[:id])
-    item.destroy
+    if @item.user_id == current_user.id
+      item = Item.find(params[:id])
+      item.destroy
+    end
     redirect_to root_path
   end
